+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Use this section to keep track of any intermediate changes that can then be move
   using only the output as inputs and ACRG repository.
 - Country mask code has now been updated to allow ocean territories (Economic Exclusion Zone, EEZ) to be included.
 - hbmcmc code now has options to use HiTRes processes for the set up
+- Add function for embedding regional fields into larger fields, in acrg.name.emissions_helper_func
 
 ### Changed
 - directory structure of acrg package has changed. Many import statements will likely need modifying